Initial custom notifications overhaul.
Implements the majority of the DLL-side functionality required for a more
flexible XML-based notifications system that can be easily trigger from
Lua or within the DLL.

Aims to provide access to notifications that open arbitrary popup windows
and have arbitrary icons.

Code is still untested. (So most likely will not work yet.)