Base Civ disables and initial tech content.
All civs from the base game, expansion, and DLC are now disabled when the
mod is loaded.

Added initial content (still missing a lot of text keys) for a big chunk
of the new technologies. Unordered and with no prereqs set. Will need to
use the Visual Tech Tree Editor to set up the new tree.