Fixed bugs relating to new SiegeMod unit abilities
Fixed a bug where the game was not correctly obtaining the unit type a civ
could construct when it was given a unit class. This caused random units
(usually settlers since the incorrect function usually returned 0) to
spawn when the free units on trade route plunder ability triggered.

Fixed a bug where the free units when surviving melee combat ability was
checking for the free units when trade route plundered XML entry.