New promotion capabilities.
Promotions can now be given free to a unit until a specific project is
completed. (Project is assumed to be global project)

Promotions can now provide a chance for a unit to attack itself.

Currently, the unit attacking itself breaks the unit graphics. May be
useful to switch to non-visualized combat for this one.