Added more main game content.
New buildings and units added to early-game.

Updated builds that still referred to old techs so they appear on the new
tree.

Updated some build feature requirements based on the removal of features
builds moving to other techs.

Added projects for cleansing saidin and the hunt for the horn.

Updated help texts for various units/buildings.